Use dask.arrays inside admm, supporting mixed sparse-dense inputs
Previously admm would rechunk the columns to be in a single chunk, and
then pass delayed numpy arrays to the local_update function.  If the
chunks along columns were of different types, like a numpy array and a
sparse array, then these would be inefficiently coerced to a single
type.

Now we pass a list of numpy arrays to the local_update function.  If
this list has more than one element then we construct a local dask.array
so that operations like dot do the right thing and call two different
local dot functions, one for each type.